pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.45.2",
    "google-generativeai>=0.8.4",
    "pillow>=11.1.0",
    "rich>=13.9.4",
]
